+++ simple_history/tests/tests/test_models.py
@@ -1128,3 +1128,22 @@ class ExtraFieldsDynamicIPAddressTestCase(TestCase):
 
         poll_history = polls[0].history.first()
         self.assertEqual('127.0.0.1', poll_history.ip_address)
+
+        
+class WarningOnAbstractModelWithInheritFalseTest(TestCase):
+    def test_warning_on_abstract_model_with_inherit_false(self):
+
+        with warnings.catch_warnings(record=True) as w:
+            class AbstractModelWithInheritFalse(models.Model):
+                string = models.CharField()
+                history = HistoricalRecords()
+
+                class Meta:
+                    abstract = True
+
+            self.assertEqual(len(w), 1)
+            self.assertTrue(issubclass(w[0].category, UserWarning))
+            self.assertEqual(str(w[0].message),
+                             'HistoricalRecords added to abstract model '
+                             '(AbstractModelWithInheritFalse) without '
+                             'inherit=True')