+++ CHANGES.rst
@@ -6,6 +6,7 @@ Unreleased
 - Add `custom_model_name` parameter to the constructor of `HistoricalRecords` (gh-451)
 - Fix header on history pages when custom site_header is used (gh-448)
 - Modify `pre_create_historircal_record` to pass `history_instance` for ease of customization (gh-421)
+- Raise warning if HistoricalRecords(inherit=False) is in an abstract model (gh-341)
 
 2.5.1 (2018-10-19)
 ------------------